Remove .onAppear modifier
1. Remove .onAppear modifier from MainView.
2. Move AuthDataResultModel to separate file.
3. Remove NavigationStack from .fullScreenCover modifier in MainView.
