+++ ReFix/ReFixApp.swift
@@ -23,7 +23,7 @@ struct ReFixApp: App {
 class AppDelegate: NSObject, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? =
                      nil) -> Bool {
-        FirebaseApp.configure ()
+        FirebaseApp.configure()
         return true
     }
 }
